huggingface_space/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart
shapely
ultralytics==8.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
pillow==10.2.0
